added arguments to rake file

+++ lib/tasks/gem_records_csv.rake
@@ -6,9 +6,14 @@ namespace :gem_records_csv do
   desc 'Loading Gem Records.'
   # Usage: cat /file/with/gem_records.txt | rake gem_records:load
 
-  task load: :environment do
+  task :load, [:filename] => :environment do |t, args|
+    filename = args[:filename]
+    if filename.nil? || filename.empty?
+      puts "Error: Filename is required"
+      exit 1
+    end
     GemRecord.delete_all
     gem_load = LoadGemRecordsCSV.new
-    gem_load.load_csv('lib/dummy_gem_records.csv')
+    gem_load.load_csv(filename)
   end
 end
